2-react-serach-agent/pyproject.toml
[project]
name = "2-react-serach-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.0",
    "langsmith>=0.7.35",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
